populate-db.py: fix module import path
At least the kamu specific 'registration' module's name clashes with a
python module with the same name. Since the python library path was
before the kamu project path, the wrong module was attempted to be
imported and we ended up with an ImportError.

To guarantee that we always include the kamu specific module move the
kamu path (and kamu's parent path) from the end of the path list to the
beginning.  Also normalize the paths for compactness.

Signed-off-by: Imre Deak <imre.deak@nokia.com>