autocomplete-combobox: fix for chrome
chrome doesn't handle keypress events the same way as firefox. At the
moment we'll get funny effects for 'l' keypresses, since its keycode
is the same as NUMPAD_ENTER. The keydown event works as expected though
and the autocomplete code uses it too to check for enter keypress, so
switch to using it.

Signed-off-by: Imre Deak <imre.deak@gmail.com>