autocomplete-combobox: emphasize matching string part in result
Note that we could also construct a single RegExp object containing all
the words to be replaced and use this RegExp object on all list items.
This would probably be faster than the approach provided by this patch,
but the following bug prevents us from using it:

Words ending or starting with a non-ascii character fail the regexp word
boundary check. So for example /\bAla-Nissilä\b/gi won't match the
"Ala-Nissilä" string.

Signed-off-by: Imre Deak <imre.deak@gmail.com>