CSS: use class instead of id to allow for later attribute redefinition
ID specific CSS rules can't be redefined by class specific rules. Since
the main column contains a lot of things for which we would want to
provide class specific rules, switch to class specific rule for the main
column too.

This is needed by an upcoming patch that wants to define a class
specific rule for an <img> within the main column.

Signed-off-by: Imre Deak <imre.deak@gmail.com>